requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary>=2.9.10
django-unfold==0.2.1
django-allauth==0.61.1
dj-rest-auth==5.0.1
drf-spectacular==0.27.1
django-cors-headers==4.3.1
Pillow>=11.0.0
python-dotenv==1.0.1
redis==5.0.4
django-redis==5.4.0
celery==5.3.6
djangorestframework-simplejwt==5.3.1